policy for post domain

+++ app/Http/Controllers/PostController.php
@@ -3,12 +3,14 @@
 namespace App\Http\Controllers;
 
 use App\Models\Post;
+use Illuminate\Support\Facades\Auth;
 use Inertia\Inertia;
 
 class PostController extends Controller
 {
     public function edit(Post $post)
     {
+        $this->authorize('update', Auth::user(), $post);
         $post->fill([
             'title' => request('title'),
             'content' => request('content'),
@@ -31,6 +33,7 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
+        $this->authorize('delete', Auth::user(), $post);
         $post->delete();
     }
 }
